python_backend/requirements.txt
fastapi==0.110.0
uvicorn==0.23.2
pydantic==2.5.2
requests==2.31.0
python-dotenv==1.0.0
python-multipart==0.0.6
firebase-admin==6.2.0
pillow==10.1.0
numpy==1.26.2
scikit-learn==1.3.2
tensorflow-hub==0.15.0
tensorflow==2.15.0
transformers==4.35.2
langchain==0.0.335
langchain-openai==0.0.2
openai==1.3.8
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
loguru==0.7.2
